requirements.txt
Flask==3.1.3
flask-cors==6.0.2
numpy==2.4.3
Pillow==12.1.1
scikit-learn==1.8.0
torch==2.10.0
torchvision==0.25.0
transformers==5.3.0
huggingface_hub==0.31.4
gunicorn==23.0.0
